+++ metrics/ris-mirror/adapter/prom/ris_mirror_prom_adapter.go
@@ -2,7 +2,7 @@ package prom
 
 import (
 	"github.com/bio-routing/bio-rd/cmd/ris-mirror/rismirror"
-	"github.com/bio-routing/bio-rd/protocols/ris/metrics"
+	"github.com/bio-routing/bio-rd/cmd/ris-mirror/rismirror/metrics"
 	"github.com/prometheus/client_golang/prometheus"
 
 	vrf_prom "github.com/bio-routing/bio-rd/metrics/vrf/adapter/prom"
@@ -11,18 +11,6 @@ import (
 const (
 	prefix = "bio_rismirror_"
 )
-
-var (
-	risMirrorSessionEstablishedDesc *prometheus.Desc
-	risMirrorObserveRIBMessages     *prometheus.Desc
-)
-
-func init() {
-	labels := []string{"sys_name", "agent_address"}
-
-	risMirrorSessionEstablishedDesc = prometheus.NewDesc(prefix+"session_established", "Indicates if a RIS session is established", labels, nil)
-	risMirrorObserveRIBMessages = prometheus.NewDesc(prefix+"observe_rib_messages", "Returns number of received rib monitoring messages", labels, nil)
-}
 
 // NewCollector creates a new collector instance for the given RIS mirror server
 func NewCollector(risMirror *rismirror.RISMirror) prometheus.Collector {
@@ -38,9 +26,6 @@ type risCollector struct {
 
 // Describe conforms to the prometheus collector interface
 func (c *risCollector) Describe(ch chan<- *prometheus.Desc) {
-	/*ch <- risMirrorSessionEstablishedDesc
-	ch <- risMirrorObserveRIBMessages*/
-
 	vrf_prom.DescribeRouter(ch)
 }
 
@@ -52,17 +37,6 @@ func (c *risCollector) Collect(ch chan<- prometheus.Metric) {
 }
 
 func (c *risCollector) collectForRouter(ch chan<- prometheus.Metric, rtr *metrics.RISMirrorRouterMetrics) {
-
-	/*l := []string{rtr.SysName, rtr.Address.String()}
-
-	ch <- prometheus.MustNewConstMetric(routeMonitoringMessagesDesc, prometheus.CounterValue, float64(rtr.RouteMonitoringMessages), l...)
-	ch <- prometheus.MustNewConstMetric(statisticsReportMessages, prometheus.CounterValue, float64(rtr.StatisticsReportMessages), l...)
-	ch <- prometheus.MustNewConstMetric(peerDownNotificationMessages, prometheus.CounterValue, float64(rtr.PeerDownNotificationMessages), l...)
-	ch <- prometheus.MustNewConstMetric(peerUpNotificationMessages, prometheus.CounterValue, float64(rtr.PeerUpNotificationMessages), l...)
-	ch <- prometheus.MustNewConstMetric(initiationMessages, prometheus.CounterValue, float64(rtr.InitiationMessages), l...)
-	ch <- prometheus.MustNewConstMetric(terminationMessages, prometheus.CounterValue, float64(rtr.TerminationMessages), l...)
-	ch <- prometheus.MustNewConstMetric(routeMirroringMessages, prometheus.CounterValue, float64(rtr.RouteMirroringMessages), l...)*/
-
 	for _, vrfMetric := range rtr.VRFMetrics {
 		vrf_prom.CollectForVRFRouter(ch, rtr.SysName, rtr.Address.String(), vrfMetric)
 	}
